Squashed 'applications/external/' changes from e43e4f6bf65..5864b3f1010
5864b3f1010 WAV Player: Fix unresponsiveness
cfea009edfb BLE Spam: Fix delay help section
bec77fd399f Format
b643a4aad2b Merge totp from https://github.com/akopachov/flipper-zero_authenticator
094d65a4018 Merge solitaire from https://github.com/doofy-dev/flipper_solitaire
4a0f5a9ee36 Merge nfc_playlist from https://github.com/acegoal07/FlipperZero_NFC_Playlist
a83d4bca832 Merge laser_tag from https://github.com/RocketGod-git/Flipper-Zero-Laser-Tag
df3c9f98b5f Merge pull request #6 from jamisonderek/jamisonderek/rfid-ammo
bb8b73ff22f Up button scans EM4100 health tag.
6e8234e7c61 Merge pull request #7 from lucaspwo/update-api-72_1
0a7ce69a2e0 Merge pull request #39 from xtruan/patch-1
c7b2006d5a2 chore: Updated version
f77e6fb59fe feat: confirm `topt export` on the device (#240) (#241)
8026dd6869d fix: add zero back
22445c50529 feat: allow delay up to 12s
25655af241b Updated to API 72.1
f0181ef71ae Update laser_tag_10px.png
48f86b96b57 Update laser_tag_10px.png
d6d4119a92a Add laser_tag from https://github.com/RocketGod-git/Flipper-Zero-Laser-Tag
54d7fe3eeb6 Fix adding new app
43ff28abf2a Merge seader from https://github.com/bettse/seader
5f1159d9604 Merge picopass from https://gitlab.com/bettse/picopass
c5e3d624357 Update for v1.1
0cece4d6f39 Update icon, remove test fap folder
2ed4e45a03c Update README.md
881263d3527 Get rid of laod_path when dealing with non-seader save
9e32cc75d0d Revert "SubGHz Bruteforcer: Fix one/two byte text"
018f9f63908 Move parse from card menu to card read success
56ac2481dd4 Debug log of save location for agnostic
7d70a7acd33 Explicitly set credential type before saving
d5a4a955da8 replace ANY_PATH references
e93f3cab71c Merge seader from https://github.com/bettse/seader
c21ec249202 Update
f9b4be203cb update
f841cb042ff updates
b2015f38869 Add contributor
6e3882f2dca Fixes
725f8891d42 Preparing for app catalog submittal
59b65a932ce Merge pull request #5 from jamisonderek/jamisonderek/always-rx
5fbc4f444dc Free IR controller if alloced.
6a56928d362 Always RX except when shooting
026c92b3994 Merge pull request #4 from jamisonderek/jamisonderek/team-color
85539e04175 Blink with team color
14067edf8c0 No github, this is not a vulnerability
7d70b9b16c1 Replace backticks with single quote marks
94f1bee7d63 Remove deprecated function
0d403739eab Update changelog
cbeb02e5d51 Bump version to 1.17
e03ad23335b Merge branch 'CVE-2024-41566' into 'main'
fb3160ca925 Merge branch 'main' of https://github.com/RocketGod-git/Flipper-Zero-Laser-Tag
0064f1e8955 Remove unused files
2d3fe78d513 Update README.md
430da4ac09b Add Game Over screen. Fix reload ammo logic.
7d2fe207529 Add beep when firing laser
0689af83f5a CVE-2024-41566: When keys are unknown emulate with a dummy MAC and ignore reader MACs
2b3aac887d9 Update README.md
4b821003009 Update README.md
f6a4170dcdf OFW version
8f00fc1199f Update game screen for extra sexiness
a31fe3f0b66 Bump to v3.1
f65e802ddd1 Correct key calculation for save as SR
cc345879c7d New splash screen and team select screen. Game screen up next.
e4251361e61 Merge pull request #3 from jamisonderek/double-stop
5bff6b7f42d Merge pull request #2 from jamisonderek/jamisonderek/ir-nec
3b05028cb92 Remove extra rx_stop call
90609a3b874 increase delay + add an address
5a6d1008678 uncompressed icons start with 0.
afde2eec5e2 Added splash screen will customize more later, added sound/vibro, added logging, still crashes after team selection and not sure why
e1d90b02450 Fixed some stuff. Added tons of logging for cli log,  but ir still crashes.
3cca0c71cfd Make the Laser Tag game.
e1fa61463e9 Initial commit

git-subtree-dir: applications/external
git-subtree-split: 5864b3f1010951a9246381caa5769b4c96053f7b

+++ .utils/add-subtree.sh
@@ -44,9 +44,13 @@ if [ "${prevremotedir}" != "" ]; then
     mv -T "${prevremotedir}" "${path}"
 fi
 
-# Add new remote at the top
-echo "${repo} ${branch} ${subdir}" | cat - "${gitsubtree}" 2> /dev/null > "${gitsubtree}.new"
-mv "${gitsubtree}.new" "${gitsubtree}"
+if [ -e "${gitsubtree}" ]; then
+    # Add new remote at the top
+    echo "${repo} ${branch} ${subdir}" | cat - "${gitsubtree}" > "${gitsubtree}.new"
+    mv "${gitsubtree}.new" "${gitsubtree}"
+else
+    echo "${repo} ${branch} ${subdir}" > "${gitsubtree}"
+fi
 git add "${gitsubtree}"
 git commit --amend --no-edit
